requirements.txt
Django==1.9
django-allauth==0.24.1
django-cors-headers==1.1.0
django-filter==0.11.0
django-rest-auth==0.6.0
djangorestframework==3.3.2
jsonfield==1.0.3
oauthlib==1.0.3
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.24.0
xmltodict==0.9.2
